moved all SQL string literals used inside tests into a single file to clean all of this up
