+++ tests/skibidi-orm/migration_engine/db_inspector/test_db_inspectors.py
@@ -8,57 +8,12 @@ from skibidi_orm.migration_engine.db_inspectors.sqlite3_inspector import (
     SQLite3Inspector,
 )
 
-sql_table1 = """
-    CREATE TABLE table1 (
-        id INTEGER PRIMARY KEY,
-        name TEXT NOT NULL
-    );
-"""
-sql_table2 = """
-    CREATE TABLE table2 (
-        id INTEGER PRIMARY KEY,
-        name TEXT NOT NULL
-    );
-"""
-sql_table_primary_key_not_null = """
-    CREATE TABLE table_primary_key_not_null (
-        id INTEGER PRIMARY KEY NOT NULL
-    );
-"""
-
-sql_schema_with_fks = [
-    """
-    CREATE TABLE users (
-        user_id INTEGER PRIMARY KEY,
-        username TEXT NOT NULL UNIQUE,
-        email TEXT NOT NULL UNIQUE,
-        password_hash TEXT NOT NULL,
-        registration_date TIMESTAMP DEFAULT CURRENT_TIMESTAMP
-    );
-""",
-    """
-    CREATE TABLE posts (
-        post_id INTEGER PRIMARY KEY,
-        user_id INTEGER NOT NULL,
-        title TEXT NOT NULL,
-        content TEXT NOT NULL,
-        post_date TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-        FOREIGN KEY (user_id) REFERENCES users(user_id)
-    );
-""",
-    """
-    CREATE TABLE comments (
-        comment_id INTEGER PRIMARY KEY,
-        username TEXT NOT NULL,
-        user_idd INTEGER NOT NULL,
-        post_id INTEGER NOT NULL,
-        comment_text TEXT NOT NULL,
-        comment_date TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-        FOREIGN KEY (user_idd, username) REFERENCES users(user_id, username),
-        FOREIGN KEY (post_id) REFERENCES posts(post_id)
-    );
-""",
-]
+from ..sql_data import (
+    sql_table1,
+    sql_table2,
+    sql_table_primary_key_not_null,
+    sql_schema_with_fks,
+)
 
 
 def execute_sqlite3_commands(db_path: str, commands: list[str]):
