+++ tests/skibidi-orm/migration_engine/state_manager/test_state_manager.py
@@ -25,67 +25,7 @@ from skibidi_orm.migration_engine.adapters.database_objects.migration_element im
 from pathlib import Path
 import pytest
 import sqlite3
-
-sql_table_user = """
-    CREATE TABLE User (
-        user_id INTEGER PRIMARY KEY,
-        user_name TEXT NOT NULL
-    );
-"""
-sql_table_post = """
-    CREATE TABLE Post (
-        post_id INTEGER PRIMARY KEY,
-        post_name TEXT NOT NULL
-    );
-"""
-
-sql_table_comment = """
-    CREATE TABLE Comment (
-        comment_id INTEGER PRIMARY KEY,
-        comment_name TEXT NOT NULL
-    );
-"""
-
-sql_table_primary_key_not_null = """
-    CREATE TABLE table_primary_key_not_null (
-        id INTEGER PRIMARY KEY NOT NULL
-    );
-"""
-
-
-# TODO: what is this?
-# sql_schema_with_fks = [
-#     """
-#     CREATE TABLE users (
-#         user_id INTEGER PRIMARY KEY,
-#         username TEXT NOT NULL UNIQUE,
-#         email TEXT NOT NULL UNIQUE,
-#         password_hash TEXT NOT NULL,
-#         registration_date TIMESTAMP DEFAULT CURRENT_TIMESTAMP
-#     );
-# """,
-#     """
-#     CREATE TABLE posts (
-#         post_id INTEGER PRIMARY KEY,
-#         user_id INTEGER NOT NULL,
-#         title TEXT NOT NULL,
-#         content TEXT NOT NULL,
-#         post_date TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-#         FOREIGN KEY (user_id) REFERENCES users(user_id)
-#     );
-# """,
-#     """
-#     CREATE TABLE comments (
-#         comment_id INTEGER PRIMARY KEY,
-#         user_idd INTEGER NOT NULL,
-#         post_id INTEGER NOT NULL,
-#         comment_text TEXT NOT NULL,
-#         comment_date TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-#         FOREIGN KEY (user_idd) REFERENCES users(user_id),
-#         FOREIGN KEY (post_id) REFERENCES posts(post_id)
-#     );
-# """,
-# ]
+from ..sql_data import sql_table_user, sql_table_post, sql_table_comment
 
 
 def execute_sqlite3_commands(db_path: str, commands: list[str]):
